forward port 1.5 fixes to 1.6

1.5 had some changes that 1.6 lacked:

- Assembly accession added to VariantStudy
- maintain soTerms API structure
- Do not wrap collections in another collection when building response
- Study type name should not be ignored (overriden via mixin)
- more tests
- Sift serializer

so 1.6 was rebased on top of 1.5. As the rebase forcefully changes the history, the original branch is left as 1.6-develop.

this PR effectively adds:
- extract dgva endpoints to its own maven submodule
- several WS refactors
- using variation-commons' Region with long start and end.

+++ eva-server/src/main/java/uk/ac/ebi/eva/server/ws/ga4gh/GA4GHBeaconWSServer.java
@@ -31,8 +31,8 @@ import uk.ac.ebi.eva.commons.core.models.VariantType;
 import uk.ac.ebi.eva.commons.core.models.ws.VariantWithSamplesAndAnnotation;
 import uk.ac.ebi.eva.commons.mongodb.services.AnnotationMetadataNotFoundException;
 import uk.ac.ebi.eva.commons.mongodb.services.VariantWithSamplesAndAnnotationsService;
-import uk.ac.ebi.eva.lib.utils.DBAdaptorConnector;
-import uk.ac.ebi.eva.lib.utils.MultiMongoDbFactory;
+import uk.ac.ebi.eva.lib.eva_utils.DBAdaptorConnector;
+import uk.ac.ebi.eva.lib.eva_utils.MultiMongoDbFactory;
 import uk.ac.ebi.eva.server.ws.EvaWSServer;
 
 import javax.servlet.http.HttpServletResponse;
@@ -53,7 +53,7 @@ public class GA4GHBeaconWSServer extends EvaWSServer {
     
     @RequestMapping(value = "/beacon", method = RequestMethod.GET)
     public GA4GHBeaconResponse beacon(@RequestParam("referenceName") String chromosome,
-                                      @RequestParam("start") int start,
+                                      @RequestParam("start") long start,
                                       @RequestParam("allele") String allele,
                                       @RequestParam("datasetIds") List<String> studies,
                                       HttpServletResponse response)
